+++ CHANGELOG.md
@@ -1,5 +1,15 @@
-Version 11.0.0
-===============================
-- Initial release for use with Airship SDK 11+
+# airship-location-cordova Plugin Changelog
+
+## Version 12.0.0 - September 3, 2020
+
+Major release updating iOS and Android Location SDK versions to 13.5.4 and 13.3.2, respectively.
+
+## Version 11.1.0 - July 15, 2019
+
+Initial release for use with Airship SDK 11+
+
+## Version 11.0.0 - July 11, 2019
+
+Initial release for use with Airship SDK 11+